Fixed problem after merge and bugs with discogLink
1. Into DefaultDiscogsService in getDiscogsLink method added transformation to lower case before comparing artist & release
2. Into DefaultDiscogsService added getDiscogsClient fot test
3. Into V1_0_initial_schema.sql added extending '.png' for junorecords
 image
4. Into vinyl.html added functionality for cross to another tab when
   click onto DoscogsLink
5. Into UsetRowMapperTest added discogsUserName
6. Into JdbcUserDaoITest added one test in order to checking if
   discogsUserName is unique
7. Into DatabasePreparerForITests & ListPreparerForTests added
   discogsUserName

// This is an in-line comment.
You can make a multi-line comment beginning with /* and ending with */. This is a multi-line comment:

/* This is a
multi-line comment */
NOTE: As you write code, you should regularly add comments to clarify the function of parts of your code. Good commenting can help communicate the intent of your code—both for others and for your future self.omments are lines of code that JavaScript will intentionally ignore. Comments are a great way to leave notes to yourself and to other people who will later need to figure out what that code does.

There are two ways to write comments in JavaScript:

Using // will tell JavaScript to ignore the remainder of the text on the current line. This is an in-line comment:

// This is an in-line comment.
You can make a multi-line comment beginning with /* and ending with */. This is a multi-line comment:

/* This is a
multi-line comment */
NOTE: As you write code, you should regularly add comments to clarify the function of parts of your code. Good commenting can help communicate the intent of your code—both for others and for your future self.